Fixed bug that was creating a form every time an object was removed or updated. Finished transition to handling a favorite item object instead of loading it from the db every time. Updated layout for the form view fragment to use the cardview that was previously added.

+++ app/src/main/java/pt/up/fe/labtablet/activities/ValidateMetadataActivity.java
@@ -188,7 +188,7 @@ public class ValidateMetadataActivity extends Activity {
 
         //remove unwanted files
         for (Descriptor desc : fItem.getMetadataItems()) {
-            if (desc.hasFile()) {
+            if (desc.hasFile() && desc.getState() == Utils.DESCRIPTOR_STATE_NOT_VALIDATED) {
                 new File(desc.getFilePath()).delete();
             }
         }
